Fix radio-group DAP issues
